requirements-full.txt
# for tools: rapid and optimal production
ortools==9.11.4210
networkx==3.1
matplotlib==3.7.5
scipy==1.10.1

# for tools: game monitor
watchdog==4.0.2
pyyaml
